requirements.txt
flask
psycopg2
pycryptodome
requests
Flask-Session